requirements.txt
opencv-python==4.8.1.78
pillow==10.0.1
tkinter
pyautogui==0.9.54
numpy==1.24.3
pynput==1.7.6
ultralytics==8.0.196
rich==13.5.2
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
pyyaml==6.0.1